// src/app/pages/master/software/model/SoftwareModel.ts
export interface SoftwareModel {
  product_code: string;
  product_name: string;
  unit: any;
  type: string;
  price: string;
}

export interface FeedbackModelSoftware {
  harga: Number;
  input_date: String;
  kode_produk: String;
  nama_produk: String;
  satuan: String;
  type: String;
  status: String;
  __v: 0;
  _id: String;
}